oylab-ai/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pdfplumber==0.11.4
python-pptx==1.0.2
pydantic==2.8.2
python-dotenv==1.0.1
supabase==2.3.5
postgrest==0.10.8
gotrue==2.4.2
storage3==0.5.3
requests==2.32.3